+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### [Unreleased][unreleased]
+#### Added
+- Added bitwise AND and XOR DSL for where clauses
+
 ### [2.1.1] - 2016-04-07
 #### Added
 - Add explicit `where(String)` method to help with non-Java language interop